+++ changelog.md
@@ -6,6 +6,7 @@
 
 - [add] 增加了 `ZhihuClient.from_url` 方法，传入合法的知乎网址，就能生成对应的对象
 - [add] 给`BaseGenerator`增加了 `add_params`和`set_params` 方法
+- [add] 写完了文档
 - [change] 改变了好多内部类名和变量名，不过对外部接口没有影响
 
 ### 0.0.2 - 2016.04.07